+++ packages/ok-components/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @ni/ok-components
 
-<!-- This log was last generated on Thu, 09 Apr 2026 19:17:36 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 15 Apr 2026 19:52:21 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.3.13
+
+Wed, 15 Apr 2026 19:52:21 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.6.0
+- Bump @ni/spright-components to v6.15.8
+
+## 0.3.12
+
+Wed, 15 Apr 2026 19:06:54 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.5.8
+- Bump @ni/spright-components to v6.15.7
+
+## 0.3.11
+
+Tue, 14 Apr 2026 20:31:57 GMT
+
+### Patches
+
+- Update fast dependency to fix view bug in Firefox ([ni/nimble@c2354bd](https://github.com/ni/nimble/commit/c2354bd5a997f90670c66a7c97d7e858607d2000))
+- Bump @ni/nimble-components to v35.5.7
+- Bump @ni/spright-components to v6.15.6
 
 ## 0.3.10